Assignment - 9/ImageNet-1k with ResNet/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorboard>=2.13.0
tqdm>=4.65.0
matplotlib>=3.7.1
albumentations>=1.3.1
pillow>=9.5.0
numpy>=1.24.3